@mcollina 2017-05-19
I think you should have an helper that spits out `key.substring(1)` if `key` starts with `@`. It should work.