@SimenB 2017-05-19
Wouldn't this be a problem for all keys not being Aa-Zz?