@SimenB 2017-05-20
I'm a bit stuck here... Why is key prefixed with `.` anyways? And is there a perf hit for doing `obj['key']` instead of `obj.key`? I've pushed what I had locally, would love some guidance on how to fix it.