@SimenB 2017-05-20
@mcollina I made it green at least. Not too happy with the implementation though... Ideas?