@SimenB 2017-05-20
Benchmarks:
Master:
```
JSON.stringify array x 3,666 ops/sec ±1.84% (81 runs sampled)
fast-json-stringify array x 2,337 ops/sec ±2.16% (83 runs sampled)
JSON.stringify long string x 14,201 ops/sec ±1.55% (80 runs sampled)
fast-json-stringify long string x 13,621 ops/sec ±1.90% (82 runs sampled)
JSON.stringify short string x 5,156,172 ops/sec ±1.87% (83 runs sampled)
fast-json-stringify short string x 12,075,130 ops/sec ±1.79% (81 runs sampled)
JSON.stringify obj x 1,729,733 ops/sec ±1.71% (79 runs sampled)
fast-json-stringify obj x 2,308,956 ops/sec ±2.22% (81 runs sampled)
```
This branch:
```
JSON.stringify array x 3,595 ops/sec ±2.19% (80 runs sampled)
fast-json-stringify array x 2,336 ops/sec ±2.11% (81 runs sampled)
JSON.stringify long string x 13,642 ops/sec ±1.65% (84 runs sampled)
fast-json-stringify long string x 13,699 ops/sec ±2.62% (84 runs sampled)
JSON.stringify short string x 4,928,457 ops/sec ±1.86% (78 runs sampled)
fast-json-stringify short string x 11,600,544 ops/sec ±1.71% (84 runs sampled)
JSON.stringify obj x 1,775,965 ops/sec ±2.34% (79 runs sampled)
fast-json-stringify obj x 2,309,468 ops/sec ±2.29% (82 runs sampled)
```