@SimenB 2017-05-20
It could have been interesting to pass the generated code through a minifier, which would remove the brackets if they're unnecessary (uglilfy does this). Not sure if you want that dep, though...