@SimenB 2017-05-20
I tested it out (because why not). Result

```
JSON.stringify array x 4,008 ops/sec ±1.64% (87 runs sampled)
fast-json-stringify array x 2,520 ops/sec ±2.12% (85 runs sampled)
JSON.stringify long string x 14,634 ops/sec ±1.31% (88 runs sampled)
fast-json-stringify long string x 14,446 ops/sec ±1.53% (85 runs sampled)
JSON.stringify short string x 5,382,246 ops/sec ±1.83% (82 runs sampled)
fast-json-stringify short string x 12,582,489 ops/sec ±2.26% (86 runs sampled)
JSON.stringify obj x 1,741,939 ops/sec ±1.57% (87 runs sampled)
fast-json-stringify obj x 2,404,493 ops/sec ±2.52% (80 runs sampled)
```

Diff:
```diff
diff --git i/index.js w/index.js
index ff80f46..d51febb 100644
--- i/index.js
+++ w/index.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const fastSafeStringify = require('fast-safe-stringify')
+const uglify = require('uglify-es')
 
 let isLong
 try {
@@ -74,10 +75,16 @@ function build (schema, options) {
     ;
     return ${main}
   `
+
+  const uglified = uglify.minify(code, {parse: {bare_returns: true}})
+
+  if (uglified.error) {
+    throw uglified.error
+  }
   if (schema.additionalProperties === true) {
-    return (new Function('fastSafeStringify', code))(fastSafeStringify)
+    return (new Function('fastSafeStringify', uglified.code))(fastSafeStringify)
   }
-  return (new Function(code))()
+  return (new Function(uglified.code))()
 }
 
 function $asNull () {
diff --git i/package.json w/package.json
index 5ebd8a8..0fb85b9 100644
--- i/package.json
+++ w/package.json
@@ -32,6 +32,7 @@
     "tap": "^10.3.0"
   },
   "dependencies": {
-    "fast-safe-stringify": "^1.1.11"
+    "fast-safe-stringify": "^1.1.11",
+    "uglify-es": "^3.0.9"
   }
 }
```