@SimenB 2017-05-20
How would we pass it through a minimizer without depending on the minimizer?