@mcollina 2017-05-20
you can try requiring it, and if it is there use it. Plus, we can put it
under an option as well.
Il giorno sab 20 mag 2017 alle 20:43 Simen Bekkhus <notifications@github.com>
ha scritto:

> How would we pass it through a minimizer without depending on the
> minimizer?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/fastify/fast-json-stringify/pull/34#issuecomment-302891281>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AADL4zT9T9oZSSRFx2DmOWKevYgWnu1Pks5r7zRjgaJpZM4NgZ7F>
> .
>