@SimenB 2017-05-20
Cool, PR sent 😄

Will you be able to merge and release this and #35 by Monday?